requirements.txt
django==3.2.15
djangorestframework==3.13.1
flake8==5.0.4
mysqlclient==2.1.0
drf-yasg==1.21.3
django-filter==22.1
